requirements.txt
django==2.0.4
gunicorn==19.7.1
whitenoise==3.3.1
coverage==4.4.1
coveralls
django-polymorphic
django-crispy-forms
django-polymorphic
